requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attention==4.0
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bert-for-tf2==0.14.7
bert4keras==0.10.6
binaryornot==0.4.4
black==20.8b1
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colabcode==0.1.0
cookiecutter==1.7.3
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
djangorestframework==3.12.2
doq==0.6.4
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Fiona==1.8.19
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gensim==3.8.3
geopandas==0.9.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
idna==2.10
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.2
jinja2-time==0.2.0
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupytext==1.6.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-bert==0.86.0
keras-contrib @ git+https://github.com/keras-team/keras-contrib@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-tcn==3.4.0
keras-transformer==0.38.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lightgbm==3.1.1
loguru==0.5.3
lxml==4.6.2
Markdown==3.3.1
markdown-it-py==0.5.6
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.3
params-flow==0.8.2
parso==0.7.1
pathspec==0.8.1
poyo==0.5.0
prometheus-client==0.9.0
protobuf==3.13.0
ptyprocess==0.6.0
py-params==0.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
Pygments==2.7.2
pylint==2.6.0
pyngrok==5.0.0
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.17.3
PySnooper==0.4.2
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.1
pyxll==5.0.6
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==5.0.3
QtPy==1.9.0
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
Send2Trash==1.5.0
sentencepiece==0.1.91
Shapely==1.7.1
six==1.15.0
sklearn==0.0
smart-open==4.2.0
soupsieve==2.0.1
sqlparse==0.4.1
tca==0.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-addons==0.11.2
tensorflow-cpu==2.3.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.10.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tokenizers==0.9.3
torch==1.7.0
tornado==6.0.4
tqdm==4.51.0
transformers==3.5.1
typed-ast==1.4.2
typeguard==2.9.1
typing-extensions==3.7.4.3
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
